% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_choropleth.R
\name{aest_choropleth}
\alias{aest_choropleth}
\title{Make a choropleth map of BC economic regions}
\usage{
aest_choropleth(tbbl, measure, region)
}
\arguments{
\item{tbbl}{a tibble that contains (at least) measure and region}

\item{measure}{the column of the tbbl that contains the measure you want to plot}

\item{region}{the column of the tbbl that contains the economic regions: these MUST be camel_case i.e. "vancouver_island_and_coast","lower_mainland_southwest","thompson_okanagan","kootenay","cariboo","north_coast_&_nechako","northeast"}
}
\description{
Make a choropleth map of BC economic regions
}
