% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_fix_labs.R
\name{aest_fix_labs}
\alias{aest_fix_labs}
\title{FUNCTION_TITLE}
\usage{
aest_fix_labs(gg)
}
\arguments{
\item{gg}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{labs}}
\code{\link[stringr]{case}},\code{\link[stringr]{str_replace}}
}
