% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_hists.R
\name{aest_hists}
\alias{aest_hists}
\title{FUNCTION_TITLE}
\usage{
aest_hists(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
