% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_level_and_change.R
\name{aest_level_and_change}
\alias{aest_level_and_change}
\title{Level and Change Plot.}
\usage{
aest_level_and_change(
  df,
  var,
  date_col,
  date_format,
  what_change = "annual",
  title = NULL,
  scale_y = "identity",
  last_n = NULL,
  caption = NULL,
  text_size = 20,
  caption_size = 12
)
}
\arguments{
\item{df}{a dataframe that contains the variable of interest, a column with a time index, and the file_name column.}

\item{var}{the column that contains the variable of interest.}

\item{date_col}{the column that contains the time index.}

\item{date_format}{the date format: see details.}

\item{what_change}{either "annual" or "from_last"}

\item{title}{Optional title to include.}

\item{scale_y}{fed into ggplot's scale_y_continuous(trans=scale_y). E.g. "log", "sqrt".}

\item{last_n}{Option to only use the last_n observations.}

\item{caption}{Over-rides the default caption (variable and file)}

\item{text_size}{font size for everything else}

\item{caption_size}{font size for caption}
}
\value{
a two panel plotly object.
}
\description{
Makes a two panel plot of the levels and changes of a non-negative variable
}
\details{
Dataframe df MUST contain a column file_name, which is shown in the plot caption.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
