% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_parts_and_whole.R
\name{aest_parts_and_whole}
\alias{aest_parts_and_whole}
\title{Plot parts and sum of parts}
\usage{
aest_parts_and_whole(
  df,
  x_var,
  y_var,
  group_var,
  position = "dodge",
  legend_spot = "right",
  caption_size = 5,
  line_colour = "black",
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{df}{dataframe that contains (at a minimum) two discrete and one numeric variables.}

\item{x_var}{a discrete variable to plot on the x-axis}

\item{y_var}{a numeric variable to be plotted on y-axis}

\item{group_var}{another discrete variable that the data is grouped by}

\item{position}{Either "stack" or "dodge", Default: 'dodge'}

\item{legend_spot}{where to put the legend, Default: 'right'}

\item{caption_size}{font size for caption, Default=5}

\item{line_colour}{the colour of the line}

\item{title}{A title, Default: NULL}

\item{subtitle}{A subtitle, Default: NULL}

\item{caption}{A caption, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
From a dataframe containing two discrete and one numeric variable,  plots bars for each combination of the discrete variables and a line for the aggregate for each value of x_var
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
