% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_plot_percent.R
\name{aest_plot_percent}
\alias{aest_plot_percent}
\title{aest_plot_percent}
\usage{
aest_plot_percent(
  df,
  x_var,
  y_var,
  filter_var,
  plot_this,
  geom,
  bcPalette = c("#1f4181", "#fabc29", "#000000", "#808080")
)
}
\arguments{
\item{df}{the dataframe that contains the data you want to plot}

\item{x_var}{the variable that will go on the x-axis... likely time.}

\item{y_var}{the thing you are plotting (could be a value, change...)}

\item{filter_var}{a column in the dataframe that you are filtering on.}

\item{plot_this}{a particular value of filter_var (the series you want to plot.)}

\item{geom}{either column or line}

\item{bcPalette}{colour palette for BC government}
}
\value{
a ggplot object
}
\description{
makes a time series plot where y variable is percent formatted
}
\details{
This function is nearly identical with aest_plot_comma.  The only difference is the formatting for the y axis and the tool tip.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
