% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_read_spreadsheet.R
\name{aest_read_spreadsheet}
\alias{aest_read_spreadsheet}
\title{wrapper function that calls one of vroom, read_csv or read_excel}
\usage{
aest_read_spreadsheet(quoted_path, quoted_file, speed = "slow", ...)
}
\arguments{
\item{quoted_path}{the (quoted) path to the file (within the project directory) e.g. "raw-data"}

\item{quoted_file}{the (quoted) name of the file}

\item{speed}{determines whether vroom or read_csv is used: read_csv is more robust, but slow, Default: 'slow'}

\item{...}{additional parameters to pass.}
}
\value{
a dataframe
}
\description{
Reads in a spreadsheet,, cleans the column names and appends a column containing the file name.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
