% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_summary.R
\name{aest_summary}
\alias{aest_summary}
\title{FUNCTION_TITLE}
\usage{
aest_summary(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{summarise}}
}
