% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_treemap.R
\name{aest_treemap}
\alias{aest_treemap}
\title{FUNCTION_TITLE}
\usage{
aest_treemap(
  df,
  group,
  value,
  limit_groups = NULL,
  show_percent = FALSE,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  lab_size = 5
)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION}

\item{value}{PARAM_DESCRIPTION}

\item{limit_groups}{PARAM_DESCRIPTION, Default: NULL}

\item{show_percent}{PARAM_DESCRIPTION, Default: FALSE}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{subtitle}{PARAM_DESCRIPTION, Default: NULL}

\item{caption}{PARAM_DESCRIPTION, Default: NULL}

\item{lab_size}{PARAM_DESCRIPTION, Default: 8}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
