% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_and_change.R
\name{parse_and_change}
\alias{parse_and_change}
\title{FUNCTION_TITLE}
\usage{
parse_and_change(df, var, date_col, date_format, what_change)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var}{PARAM_DESCRIPTION}

\item{date_col}{PARAM_DESCRIPTION}

\item{date_format}{PARAM_DESCRIPTION}

\item{what_change}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[lubridate]{interval}}
}
