% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_level_and_change.R
\name{plot_level_and_change}
\alias{plot_level_and_change}
\title{FUNCTION_TITLE}
\usage{
plot_level_and_change(df, var_name, title, scale_y, caption, caption_size)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}

\item{var_name}{PARAM_DESCRIPTION}

\item{title}{PARAM_DESCRIPTION}

\item{scale_y}{PARAM_DESCRIPTION}

\item{caption}{PARAM_DESCRIPTION}

\item{caption_size}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
