% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_choropleth.R
\name{aest_choropleth}
\alias{aest_choropleth}
\title{Make a choropleth map of BC economic regions}
\usage{
aest_choropleth(tbbl, region, thingy, value)
}
\arguments{
\item{tbbl}{a tibble that contains (at least) the following 3 columns:}

\item{region}{the column of the tbbl that contains the economic regions: these MUST be camel_case i.e. mainland_south_west, vancouver_island_coast, north_coast_&_nechako, cariboo, kootenay, north_east, thompson_okanagan}

\item{thingy}{the column of the tbbl that contains the name of what is being plotted.}

\item{value}{the column of the tbble that contains the value of what is being plotted.}
}
\description{
Make a choropleth map of BC economic regions
}
