% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aest_treemap.R
\name{aest_treemap}
\alias{aest_treemap}
\title{FUNCTION_TITLE}
\usage{
aest_treemap(
  df,
  grp,
  vrbl,
  limit_groups = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  lab_size = 5,
  caption_size = 5,
  legend_spot = "right"
)
}
\arguments{
\item{df}{A dataframe that must contain column file_name (for auto caption)}

\item{grp}{a grouping variable}

\item{vrbl}{the variable being plotted}

\item{limit_groups}{what is the maximum number of groups to consider, Default: NULL}

\item{title}{title if you want, Default: NULL}

\item{subtitle}{subtitle if you want, Default: NULL}

\item{caption}{override the default caption, Default: NULL}

\item{lab_size}{blunt control over labeling size, Default: 5}

\item{caption_size}{font size for caption Default: 5}

\item{legend_spot}{where to put the legend Default: "right"}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
